pyproject.toml
[project]
name = "his-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "prometheus-client>=0.23.1",
    "psutil>=7.1.3",
    "python-logging-loki>=0.3.1",
    "requests>=2.32.5",
]
